Moved accumulating_ring_buffer into detection header
